requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.19.0
opencv-python>=4.5.0
Pillow>=8.0.0
PyYAML>=5.4.0
easydict>=1.9
tqdm>=4.60.0
xlsxwriter>=3.0.0
